Decouple weighted pool tests a bit
